% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs2pc_dataset.R
\docType{methods}
\name{bbs2pc_dataset}
\alias{bbs2pc_dataset}
\title{create BBS point Count table based on BBS Fifty}
\usage{
bbs2pc_dataset(xytbl, lookupTables)
}
\arguments{
\item{xytbl}{A path to access current XY table.}

\item{lookupTables}{A list of path to access lookupTables.}
}
\value{
Vector of available url to download
}
\description{
The function produce a character vector of file available to download. It uses a preset table
in which a list of relevant dataset name, their associate url and password are stored. To
retrive available file, the function derive URL, and username/password using the \code{datasetName}
}
\examples{
url <- "ftp://ftpext.usgs.gov/pub/er/md/laurel/BBS/DataFiles/50-StopData/1997ToPresent_SurveyWide/"
bbs2pc_dataset(xytbl, lookupTables)
}
\author{
Melina Houle
}
