% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs2pc_detection.R
\docType{methods}
\name{bbs2pc_detection}
\alias{bbs2pc_detection}
\title{create BBS point Count table based on BBS Fifty}
\usage{
bbs2pc_detection(
  fiftyDir,
  datasetName,
  lookupTables,
  minYear = 1965,
  maxYear = 2019,
  survey_tbl,
  dropZeros = TRUE
)
}
\arguments{
\item{fiftyDir}{A character string giving the path where the downloaded files are saved.}

\item{datasetName}{A character string. Name of the dataset from which BBS PointCount table will be created. Default is all fifty files.}

\item{lookupTables}{A list of path to access lookupTables.}

\item{minYear}{Numerical giving the lower bound of year to be included. Default is \code{1965}.}

\item{maxYear}{Numerical giving the upper bound of year to be included. Default is \code{2019}.}

\item{survey_tbl}{An dataframe from which geographical range is subset and visit foreign_key is derived.}

\item{dropZeros}{Logical Drop all species with abundance being zero. Default is TRUE.}
}
\value{
Vector of available url to download
}
\description{
The function produce a character vector of file available to download. It uses a preset table
in which a list of relevant dataset name, their associate url and password are stored. To
retrive available file, the function derive URL, and username/password using the \code{datasetName}
}
\examples{
fiftyDir = outDir
lookupTables = list(regionCodes = file.path(lookupDir,fregionCodes), 
                    speciesList = file.path(lookupDir,fSpecies)) 
example <-bbs2pc_detection(fiftyDir, c("fifty1.csv"), lookupTables, minYear=2018, maxYear=2019, bbs_visit)
}
\author{
Melina Houle
}
