% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs2pc_location.R
\docType{methods}
\name{bbs2pc_location}
\alias{bbs2pc_location}
\title{Create XY table}
\usage{
bbs2pc_location(xytbl, lookupTables, dataset_tbl)
}
\arguments{
\item{xytbl}{A path to access current XY table.}

\item{lookupTables}{A list of path to access lookupTables.}

\item{dataset_tbl}{A dataframe with dataset parameters.}
}
\value{
XY table
}
\description{
The function used SScombo received from Diana Stralberg and create XY table according to coordinate we have
}
\examples{
bbs2pc_location(xytbl, lookupTables, dataset_tbl)
}
\author{
Melina Houle
}
