% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbs2pc_survey.R
\docType{methods}
\name{bbs2pc_survey}
\alias{bbs2pc_survey}
\title{Convert BBS weather file into BAM PKEY table}
\usage{
bbs2pc_survey(
  fweather,
  lookupTables,
  minYear = 1965,
  maxYear = 2019,
  pc_location
)
}
\arguments{
\item{fweather}{A character string giving the path where the downloaded file is saved.}

\item{lookupTables}{A list of path to access lookupTables.}

\item{minYear}{Numerical giving the lower bound of year to be included. Default is \code{1965}.}

\item{maxYear}{Numerical giving the upper bound of year to be included. Default is \code{2019}.}

\item{pc_location}{A location table from which foreign_key is derived.}
}
\value{
Data frame representing PKEY table
}
\description{
The function uses weather.csv and to create BBS PKEY table according to the BAM standard
}
\examples{
outDir <- "C:/data"
bbs2pc_survey(outDir, lookupTables,minYear=1965, maxYear=2019, pc_location)
}
\author{
Melina Houle
}
