% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwdBBS.R
\docType{methods}
\name{dwdBBS}
\alias{dwdBBS}
\title{Download BBS Fifty from ftp site.}
\usage{
dwdBBS(urls, urlData, datasetName, outDir, extract = FALSE)
}
\arguments{
\item{urls}{A character string. Represents the url of file to be downloaded.}

\item{urlData}{A character string. Representing the URL to access the zipped Fifty.}

\item{datasetName}{A character string. Name of the dataset to download.}

\item{outDir}{A character string giving the path where the downloaded file is saved.}

\item{extract}{Logical. If \code{TRUE}, file is untar and/or unzip. Default is \code{FALSE}.}
}
\value{
Invoked for its side-effect of downloading files to the \code{destfile/} directory.
}
\description{
To avoid downloading existing files, the function verify if file exists locally prior to download.
When file exists, the function compare checksum value from checksum logged from original download.
}
\examples{
url <- "http://ftp.geogratis.gc.ca/pub/nrcan_rncan/archive/vector/cli_itc_50k/land_use"
dataList <- "land_use/L040J03.zip"
dwdBBS(url = url, destfile = tempdir(), extract = FALSE)
}
\author{
Melina Houle
}
